Chore: 산 정보 컴포넌트 공용 컴포넌트로 분리, 산 상세 및 모임 페이지 지도 연동
